{{cookiecutter.project_slug}}/requirements/local.txt
-r base.txt

django-extensions==3.2.3
factory-boy==3.3.0
Faker==24.0.0
pytest==8.0.2
pytest-django==4.8.0
pytest-factoryboy==2.6.1
python-dateutil==2.9.0.post0